feat(widget): updated UI to handle no tags case
